requirements.txt
pynput==1.7.6
numpy==1.24.3
Pillow==9.5.0
mss==9.0.1
